separate UploadView into fileupload dir, still need to fix dashboard.uploadhander issue
