+++ cadetapp/dashboard/templates/dashboard/file_upload.html
@@ -3,7 +3,7 @@
 <!-- Breadcrumbs-->
 <ol class="breadcrumb">
   <li class="breadcrumb-item">
-    <a href="/dashboard">Dashboard</a>
+    <a href="/">Dashboard</a>
   </li>
   <li class="breadcrumb-item active"> Upload</li>
 </ol>
